requirements.txt
gym>=0.21.0
stable-baselines3==1.8.0
torch>=1.10.0
numpy>=1.20.0
networkx>=2.6.0
matplotlib>=3.5.0
pyyaml>=6.0.0
